Build improvements, compiler warning fixes and build fixes, and lots of clean up.
* Created separate SVN version header.
* Add true config.h for platform dependency checks.
* Add dependency check cmake module.
* Remove some leftover hand crafted make files.
* Remove non-cmake build instructions from COMPILING.txt.
* Fix split _() strings causing Visual C++ compiler error.
* Fix lots of compiler warnings.
* Change project file parameter container from wxArray to boost::vector_ptr.
* Removed lots of redundant header definitions.
* Fixed green_xpm redefinition in ercgreen.xpm.
* Remove some dead code and unnecessary class methods.

+++ eeschema/locate.cpp
@@ -733,7 +733,7 @@ LibEDA_BaseStruct* LocateDrawItem( SCH_SCREEN*             Screen,
                 break;
             dx = aRefPoint.x - Circle->m_Pos.x;
             dy = aRefPoint.y + Circle->m_Pos.y;
-            int dist = (int) sqrt( dx * dx + dy * dy );
+            int dist = (int) sqrt( (double)( dx * dx + dy * dy ) );
             if( abs( dist - Circle->m_Rayon ) <= seuil )
                 return DrawItem;
         }
